requirements.txt
httpx>=0.27.0
streamlit>=1.35.0
playwright>=1.44.0
beautifulsoup4>=4.12.3
pandas>=2.2.3
openpyxl>=3.1.2
fastapi>=0.111.0
uvicorn>=0.30.1
python-dotenv>=1.0.1
